--- src/main/java/com/spiro/operation/dto/StationRequest.java
@@ -1,4 +1,0 @@
-package com.spiro.operation.dto;
-
-public class StationRequest {
-}
