--- src/main/java/com/spiro/operation/dto/StationResponse.java
@@ -1,4 +1,0 @@
-package com.spiro.operation.dto;
-
-public class StationResponse {
-}
