--- src/main/java/com/spiro/operation/model/IncidentType.java
@@ -1,11 +1,0 @@
-
-package com.spiro.operation.model;
-
-public enum IncidentType {
-    BATTERY_ISSUE,
-    VEHICLE_ISSUE,
-    CHARGING_ISSUE,
-    PAYMENT_ISSUE
-
-
-}
