--- src/main/java/com/spiro/operation/model/ShiftRecords.java
@@ -1,70 +1,0 @@
-package com.spiro.operation.model;
-
-import java.time.LocalDateTime;
-
-public class ShiftRecords {
-
-    private String agentId;
-    private String StationId;
-    private LocalDateTime timeIn;
-    private LocalDateTime timeOut;
-    private String location;
-    private int units;
-    private String comments;
-
-    public String getAgentId() {
-        return agentId;
-    }
-
-    public void setAgentId(String agentId) {
-        this.agentId = agentId;
-    }
-
-    public String getStationId() {
-        return StationId;
-    }
-
-    public void setStationId(String stationId) {
-        StationId = stationId;
-    }
-
-    public LocalDateTime getTimeIn() {
-        return timeIn;
-    }
-
-    public void setTimeIn(LocalDateTime timeIn) {
-        this.timeIn = timeIn;
-    }
-
-    public LocalDateTime getTimeOut() {
-        return timeOut;
-    }
-
-    public void setTimeOut(LocalDateTime timeOut) {
-        this.timeOut = timeOut;
-    }
-
-    public String getLocation() {
-        return location;
-    }
-
-    public void setLocation(String location) {
-        this.location = location;
-    }
-
-    public int getUnits() {
-        return units;
-    }
-
-    public void setUnits(int units) {
-        this.units = units;
-    }
-
-    public String getComments() {
-        return comments;
-    }
-
-    public void setComments(String comments) {
-        this.comments = comments;
-    }
-}
