--- src/main/java/com/spiro/operation/model/ShiftType.java
@@ -1,8 +1,0 @@
-package com.spiro.operation.model;
-
-public enum ShiftType {
-    HOURLY,
-    DAILY,
-    WEEKLY,
-    MONTHLY
-}
