--- src/main/java/com/spiro/operation/model/StationType.java
@@ -1,7 +1,0 @@
-package com.spiro.operation.model;
-
-public enum StationType {
-    BRICKANDMORTAR,
-    AUTOMATED,
-    INTERGRATED
-}
