--- src/main/java/com/spiro/operation/model/Status.java
@@ -1,6 +1,0 @@
-package com.spiro.operation.model;
-
-public class Status {
-
-
-}
