--- src/main/java/com/spiro/operation/repository/AttendantRepository.java
@@ -1,14 +1,0 @@
-package com.spiro.operation.repository;
-
-
-import com.spiro.operation.model.Attendants;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface AttendantRepository extends JpaRepository<Attendants, String> {
-    List<Attendants> findByfirstName(String Attendants);
-
-}
