--- src/main/java/com/spiro/operation/repository/StationRepository.java
@@ -1,12 +1,0 @@
-package com.spiro.operation.repository;
-
-import com.spiro.operation.model.Station;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface StationRepository extends JpaRepository {
-    Station save(List<Station> stations);
-
-    Station findAll(List<Station> stations);
-}
