+++ src/main/resources/application.properties
@@ -1,8 +1,11 @@
 spring.application.name=operation
 
-spring.datasource.url=jdbc:postgresql://localhost:5050/spiro_records_db
-spring.datasource.username=spiro_user
-spring.datasource.password=spiro123
+spring.datasource.url=jdbc:postgresql://localhost:5432/Operations
+spring.datasource.username=admin
+spring.datasource.password=2002
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
